+++ src/StateSmith/Output/UserConfig/AutoVars/PythonAutoVarsParser.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System;
 using System.Text.RegularExpressions;
 
-namespace StateSmith.Output.UserConfig;
+namespace StateSmith.Output.UserConfig.AutoVars;
 
 public class PythonAutoVarsParser : IAutoVarsParser
 {