+++ src/StateSmith/SmGraph/PreDiagramSettingsReader.cs
@@ -1,6 +1,7 @@
 #nullable enable
 
 using StateSmith.Output.UserConfig;
+using StateSmith.Output.UserConfig.AutoVars;
 using StateSmith.Runner;
 
 namespace StateSmith.SmGraph;