+++ src/StateSmithTest/Output/Algos/AlgoTests.cs
@@ -27,7 +27,7 @@ public class AlgoTests
     }
 
     [Fact]
-    public void IntegrationTests_SingleEvent_C99_CSharp_Javascript_Java_Python()
+    public void IntegrationTests_SingleEvent_C99_CSharp_Javascript_Java_Python_TypeScript()
     {
         const string basicSingleEventPlantUml = """
             @startuml RocketSm
@@ -44,27 +44,9 @@ public class AlgoTests
     }
 
     [Fact]
-    public void IntegrationTests_MultipleEvent_C99_CSharp_Javascript_Java_Python()
+    public void IntegrationTests_MultipleEvent_C99_CSharp_Javascript_Java_Python_TypeScript()
     {
-        const string basicSingleEventPlantUml = """
-            @startuml RocketSm
-            [*] --> group
-            state s1
-            
-            state group {
-                state g1
-                state g2
-                [*] --> g1
-            }
-
-            g1 --> g2: EV1 [a > 20]
-            g2 --> g1: EV2
-            group --> s1: EV1
-            
-            @enduml
-            """;
-
-        RunIntegrationTestMatrix(basicSingleEventPlantUml, "out2");
+        RunIntegrationTestMatrix(allLangsMultipleEvents, "out2");
     }
 
     private static void RunIntegrationTestMatrix(string MinimalPlantUmlFsm, string outDirName)
@@ -85,6 +67,10 @@ public class AlgoTests
 
                 // Python only supports Balanced2 right now. See https://github.com/StateSmith/StateSmith/issues/398
                 if (transpilerId == TranspilerId.Python && algoId != AlgorithmId.Balanced2)
+                    continue;
+
+                // TypeScript only supports Balanced2 right now. See https://github.com/StateSmith/StateSmith/issues/407
+                if (transpilerId == TranspilerId.TypeScript && algoId != AlgorithmId.Balanced2)
                     continue;
 
                 var dirName = $"{outDir}/{algoId}_{transpilerId}";
@@ -109,5 +95,68 @@ public class AlgoTests
     {
         return Enum.GetValues(typeof(T)).Cast<T>();
     }
+
+    const string allLangsMultipleEvents = """"
+        @startuml RocketSm
+        [*] --> group
+        state s1
+            
+        state group {
+            state g1
+            state g2
+            [*] --> g1
+        }
+
+        g1 --> g2: EV1 [a > 20]
+        g2 --> g1: EV2
+        group --> s1: EV1
+        
+        /'! $CONFIG: toml
+
+            [RenderConfig.C]
+            CFileTop = """
+                // RenderConfig.C.CFileTop
+                """
+
+            [RenderConfig.CSharp]
+            ClassCode = """
+                public void UserCodeMethod() {
+                    Console.WriteLine('userCodeMethod');
+                }
+                """
+
+            [RenderConfig.Java]
+            ClassCode = """
+                public void userCodeMethod() {
+                    System.out.println('userCodeMethod');
+                }
+                """
+
+            [RenderConfig.Python]
+            ClassCode = """
+                def userCodeMethod(self):
+                    print('userCodeMethod')
+                """
+
+            [RenderConfig.JavaScript]
+            ClassCode = """
+                function userCodeMethod() {
+                    console.log('userCodeMethod');
+                }
+                """
+
+            [RenderConfig.TypeScript]
+            ClassCode = """
+                public userCodeMethod() {
+                    console.log('userCodeMethod');
+                }
+                """
+            Implements = "IRocketSm"
+            Extends = "RocketSmBase"
+
+        '/
+        @enduml
+        """";
+
 }